refactor: circular dependency
