+++ packages/core/src/extensionShared.ts
@@ -42,8 +42,8 @@ import { UriHandler } from './shared/vscode/uriHandler'
 import { disableAwsSdkWarning } from './shared/awsClientBuilder'
 import { FileResourceFetcher } from './shared/resourcefetcher/fileResourceFetcher'
 import { ResourceFetcher } from './shared/resourcefetcher/resourcefetcher'
-import { ExtStartUpSources, getAuthFormIdsFromConnection } from './shared/telemetry/util'
-import { ExtensionUse } from './auth/utils'
+import { ExtStartUpSources } from './shared/telemetry/util'
+import { ExtensionUse, getAuthFormIdsFromConnection } from './auth/utils'
 import { Auth } from './auth'
 import { AuthFormId } from './auth/ui/vue/authForms/types'
